jsonpath: add scalar evaluation
This commit adds functionality for scalar evalation for jsonpath
queries, and adds floats as scalars.

Epic: None
Release note: None
