+++ pkg/util/jsonpath/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "jsonpath",
-    srcs = ["expr.go"],
+    srcs = [
+        "expr.go",
+        "scalar.go",
+    ],
     importpath = "github.com/cockroachdb/cockroach/pkg/util/jsonpath",
     visibility = ["//visibility:public"],
     deps = [
