+++ pkg/util/jsonpath/parser/testdata/jsonpath
@@ -252,6 +252,16 @@ $.a[$.a]
 ----
 $."a"[$."a"] -- normalized!
 
+parse
+$[true]
+----
+$[true]
+
+parse
+$[false]
+----
+$[false]
+
 # postgres allows floats as array indexes
 # parse
 # $.abc[1.0]
@@ -285,14 +295,6 @@ $."a"[$."a"] -- normalized!
 
 # parse
 # $.1a
-# ----
-
-# parse
-# $[true]
-# ----
-
-# parse
-# $[false]
 # ----
 
 # parse
